Rename db_transformations to identifier_utils
