+++ f/common_logic/identifier_utils.script.yaml
@@ -1,6 +1,6 @@
-summary: 'Shared logic for database transformations'
+summary: 'Shared logic for transforming and sanitizing identifiers'
 description: ''
-lock: '!inline f/common_logic/db_transformations.script.lock'
+lock: '!inline f/common_logic/identifier_utils.script.lock'
 has_preprocessor: false
 kind: script
 no_main_func: true
