+++ tox.ini
@@ -48,8 +48,8 @@ commands =
 [testenv:common_logic]
 deps =
     -r{toxinidir}/f/common_logic/data_conversion.script.lock
-    -r{toxinidir}/f/common_logic/db_transformations.script.lock
     -r{toxinidir}/f/common_logic/db_operations.script.lock
+    -r{toxinidir}/f/common_logic/identifier_utils.script.lock
     -r{toxinidir}/f/common_logic/file_operations.script.lock
     -r{toxinidir}/f/common_logic/tests/requirements-test.txt
 commands =
